TheNetNinja tutorial >> FINISHED PROJECT

+++ routes/web.php
@@ -33,8 +33,18 @@ Route::get('/', function () {
 
 
 //The actions are in the Http >> Controllers >> PizzaController.php (the second reference tells the controller and the exact action)
-Route::get('/pizzas', 'PizzaController@index');
-Route::get('/pizzas/create', 'PizzaController@create');     //You have to put this route before the {id} one, becaouse if not, the create word will be the ID --> wrong view
-Route::post('/pizzas', 'PizzaController@store');
-Route::get('/pizzas/{id}', 'PizzaController@show');
-Route::delete('/pizzas/{id}', 'PizzaController@destroy');
+                                                     //pizzas. >> namespace, pl hambiknál hamburger. lesz
+Route::get('/pizzas', 'PizzaController@index')->name('pizzas.index')->middleware('auth');
+Route::get('/pizzas/create', 'PizzaController@create')->name('pizzas.create');     //You have to put this route before the {id} one, becaouse if not, the create word will be the ID --> wrong view
+Route::post('/pizzas', 'PizzaController@store')->name('pizzas.store');
+Route::get('/pizzas/{id}', 'PizzaController@show')->name('pizzas.show')->middleware('auth');
+Route::delete('/pizzas/{id}', 'PizzaController@destroy')->name('pizzas.destroy')->middleware('auth');
+
+//UI package generálta
+//ezen a weboldalon csak a staffok tudnak regisztrálni
+Auth::routes([
+    'register' => false     //ha már regisztráltál, nem tudsz mégegyszer
+]);
+//General: Auth::routes();
+
+Route::get('/home', 'HomeController@index')->name('home');
